+++ Automation/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v2.7.2 (pre-release, 10/6/2025):
 * [Fix] Properly display "add" operator when it has more than 2 arguments.
+* [Change] Improve script errors handling in UI. Fix some edge cases when game could crash due to a script error.
 
 # v2.7.1 (September 5th, 2025):
 * [Fix] Fix the "can plant" template definition.